product_module/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
beautifulsoup4==4.11.1
boto==2.49.0
boto3==1.22.5
botocore==1.25.5
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
class-registry==2.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.50.0
django-environ==0.8.1
django-graph-api==0.3.0
django-rest-auth==0.9.5
django-s3direct==2.0.0
django-storages==1.12.3
djangorestframework==3.11.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
google==3.0.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
graphql-py==0.8.1
httpie==3.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
ply==3.11
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==2.74.0
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
